+++ cmd/gaiad/cmd/root_test.go
@@ -4,14 +4,13 @@ import (
 	"testing"
 
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
-
-	app "github.com/cosmos/gaia/v6/app"
-	"github.com/cosmos/gaia/v6/cmd/gaiad/cmd"
 	"github.com/stretchr/testify/require"
+
+	app "github.com/cosmos/gaia/v8/app"
+	"github.com/cosmos/gaia/v8/cmd/gaiad/cmd"
 )
 
 func TestRootCmdConfig(t *testing.T) {
-
 	rootCmd, _ := cmd.NewRootCmd()
 	rootCmd.SetArgs([]string{
 		"config",          // Test the config cmd
@@ -19,5 +18,5 @@ func TestRootCmdConfig(t *testing.T) {
 		"test",            // value
 	})
 
-	require.NoError(t, svrcmd.Execute(rootCmd, app.DefaultNodeHome))
+	require.NoError(t, svrcmd.Execute(rootCmd, "", app.DefaultNodeHome))
 }