+++ cmd/gaiad/main.go
@@ -6,14 +6,14 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 
-	app "github.com/cosmos/gaia/v6/app"
-	"github.com/cosmos/gaia/v6/cmd/gaiad/cmd"
+	app "github.com/cosmos/gaia/v8/app"
+	"github.com/cosmos/gaia/v8/cmd/gaiad/cmd"
 )
 
 func main() {
 	rootCmd, _ := cmd.NewRootCmd()
 
-	if err := svrcmd.Execute(rootCmd, app.DefaultNodeHome); err != nil {
+	if err := svrcmd.Execute(rootCmd, "", app.DefaultNodeHome); err != nil {
 		switch e := err.(type) {
 		case server.ErrorCode:
 			os.Exit(e.Code)