+++ tests/e2e/e2e_util_test.go
@@ -7,8 +7,10 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"strings"
 	"time"
 
+	"github.com/cosmos/cosmos-sdk/client/flags"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	"github.com/ory/dockertest/v3/docker"
@@ -61,6 +63,66 @@ func (s *IntegrationTestSuite) connectIBCChains() {
 	)
 
 	s.T().Logf("connected %s and %s chains via IBC", s.chainA.id, s.chainB.id)
+}
+
+func (s *IntegrationTestSuite) sendMsgSend(c *chain, valIdx int, from, to, amt, fees string) {
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+
+	s.T().Logf("sending %s tokens from %s to %s on chain %s", amt, from, to, c.id)
+
+	exec, err := s.dkrPool.Client.CreateExec(docker.CreateExecOptions{
+		Context:      ctx,
+		AttachStdout: true,
+		AttachStderr: true,
+		Container:    s.valResources[c.id][valIdx].Container.ID,
+		User:         "root",
+		Cmd: []string{
+			"gaiad",
+			"tx",
+			"bank",
+			"send",
+			from,
+			to,
+			amt,
+			fmt.Sprintf("--%s=%s", flags.FlagChainID, c.id),
+			fmt.Sprintf("--%s=%s", flags.FlagFees, fees),
+			"--keyring-backend=test",
+			"--broadcast-mode=sync",
+			"--output=json",
+			"-y",
+		},
+	})
+	s.Require().NoError(err)
+
+	var (
+		outBuf bytes.Buffer
+		errBuf bytes.Buffer
+	)
+
+	err = s.dkrPool.Client.StartExec(exec.ID, docker.StartExecOptions{
+		Context:      ctx,
+		Detach:       false,
+		OutputStream: &outBuf,
+		ErrorStream:  &errBuf,
+	})
+	s.Require().NoErrorf(err, "stdout: %s, stderr: %s", outBuf.String(), errBuf.String())
+
+	var txResp sdk.TxResponse
+	s.Require().NoError(cdc.UnmarshalJSON(outBuf.Bytes(), &txResp))
+
+	endpoint := fmt.Sprintf("http://%s", s.valResources[c.id][valIdx].GetHostPort("1317/tcp"))
+
+	// wait for the tx to be committed on chain
+	s.Require().Eventuallyf(
+		func() bool {
+			return queryGaiaTx(endpoint, txResp.TxHash) == nil
+		},
+		time.Minute,
+		5*time.Second,
+		"stdout: %s, stderr: %s",
+		outBuf.String(), errBuf.String(),
+	)
 }
 
 func (s *IntegrationTestSuite) sendIBC(srcChainID, dstChainID, recipient string, token sdk.Coin) {
@@ -134,6 +196,20 @@ func queryGaiaTx(endpoint, txHash string) error {
 	}
 
 	return nil
+}
+
+func getSpecificBalance(endpoint, addr, denom string) (amt sdk.Coin, err error) {
+	balances, err := queryGaiaAllBalances(endpoint, addr)
+	if err != nil {
+		return amt, err
+	}
+	for _, c := range balances {
+		if strings.Contains(c.Denom, denom) {
+			amt = c
+			break
+		}
+	}
+	return
 }
 
 func queryGaiaAllBalances(endpoint, addr string) (sdk.Coins, error) {